requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64

tensorflow-gpu=1.15.3
scikit-learn=0.22.2
numpy=1.18.4
pandas=1.0.4
matplotlib=3.1.3